requirements.txt
gradio
matplotlib
pillow
seaborn